requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
python-jose
python-multipart
cryptography
boto3
psycopg2
python-multipart
pillow
fastapi-mail
fastapi-utils
PyMySQL
alembic
pydantic
requests
xendit-python